Fixed IDEA-59641 "Git: externally commited file isn't unmarked until explicit refresh"
Introduced GitIndexChangeListener to listen to .git/index of all repositories in the project.
To avoid marking dirty, when IDEA touches .git/index (while committing, adding or else) notify GitIndexChangeListener when IDEA's operation starts and when it ends. Notification is made from GitHandler.start; several GitCommands are flagged as index modifiers.