IDEA-64523 remove auth info from hgrc after clone
Currently hg4idea doesn't work with mercurial prompt - instead for tasks requiring authentication it asks it from user and then injects into the URL, which is in turn supplied to an hg command (push, pull, clone, etc.).
Because of that after cloning authentication data was recorded to .hg/hgrc to paths.default, which unsecure.

Now after clone .hg/hgrc is scanned for paths.default, and user info data is removed from there.
It is a hotfix. Right fix is to use hg prompt to supply password, will be implemented later.