IDEA-64895 GitPushActiveBranches view button
View button (to view commit details) produces a request to git, thus it is made in Task.Backgroundable.
It was run in the same BackgroundTaskQueue as fetching, and thus was block by the fetching process.
It is not necessary to use the same queue since the view task executes read operations only.