GitMergeConflictResolver: don't auto-commit from notification.
If user updates, gets merge conflict, resolves it at once, - it is natural to continue update.
But if user explicitly cancelled resolving merge conflict and then reinvoked the Merge Tool again from notification, don't autocommit or autocontinue rebase, because it may surprise the user.

For this introduce justMerge method which doesn't proceed with operation.