GitRepositoryManager: update on .git dir change, not on mappings change
This solves IDEA-64973 ("Enable Git Integration" => mapping is changed => GitRepositoryManager gets notified before git init happens => fails to create a GitRepository and shows a notification).
This also handles the case when a registered git repository was created or deleted outside IDEA - now it will be captured.

updateRepositoriesCollection() is called on startup now to initially collect the repositories.
For this GitRepositoryManager.getInstance() is removed from GitUntrackedFilesHolder constructor (otherwise it leads to a cyclic dependency).

Git root rescan happens on every updateRepositoriesCollection().