Fix 'git fetch --prune' deleting all remote refs.
Before Git 1.7.7.2 'git fetch --prune origin master' used to remove all remote references.
This was fixed in 1.7.7.2.
For the sake of simplicity we are not going to call '--prune' for the older Gits.

Other way would be to call 'git remote prune origin' for older Gits,
but this auto-prune feature doesn't seem to be important enough for such effort.