IDEA-83604 Deleting branch unmerged to upstream
If branch is merged to the HEAD, but not merged to the upstream, it also can't be deleted by 'git branch -d'.
Capture this case with additional detector, and use baseBranch parameter (which is either current branch or the upstream branch) to make the list of unmerged commits.
The pattern is checked for 1.7.0.0 and 1.7.9.2
Write a test for this situation.

(cherry picked from commit 85429d7)
[reviewed by irengrig]