Git.runInBackground: don't use the queue.
The most of Git activity is produce by commands invoked by the user.
In this case there is no sense in making the user wait until the previous command is completed
(for example, fetch can take a very long time, why should I wait for it to see the history of a file?)
Most of the others (e.g. in GitChangeProvider or GitUpdateProcess) are executed in their own general VCS processes.
The write-lock is taken anyway globally in GitHandler when needed.