[git] deprecate GitUtil.getGitRoot... methods, change in GitChangeProvider
Ground:
GitUtil.getGitRootOrNull is expensive, because it walks through a file system, which may be a hotspot when called from GitChangesCollector with thousand of dirty files.
There is no need to get the actual .git root: we can get the VCS root (defined in the Settings), because we notify user about Git roots problems anyway.

 Changes:
 Mark GitUtil.getGitRoot... methods as deprecated.
 Change the code in GitChangesCollector to use ProjectLevelVcsManager#getVcsRootFor.