[git] IDEA-87609 When updating via Rebase, use merge instead of rebase, if it is fast-forward.
Before starting the normal update procedure (with possible stash-unstash), try to perform a fast-forward merge instead of rebase, because it is equal to fast-forward rebase, but is faster, because doesn't require stash-unstash (unless the uncommitted files would be overwritten by merge)
Any failure in 'git merge --ff-only' means, than this repository will then be updated in a usual manner.
No pre-checking if a fast-forward merge is possible is done, because it is fast enough to execute the command and get the error.