[git] Fix show diff with parent for merge revisions, especially for moved files (IDEA-89376)
For merge revisions collect correct information about file paths of the file, how it was named in both parent revisions.
For it call git diff -M --name-status <parent>..<current> for both parents: if the file was renamed in one of merged branches, it will be noted in the diff output.
This process happens in the background, along with check if the file was touched in the revision.
Btw, use correct path to check if the file was touched in the merge revision (the file can be renamed after this merge, and current name can be different).