IDEA-86065 Don't treat the registered Git root as invalid if it was not auto-detected
The GitRootDetector for the sake of performance doesn't detect Git roots deeper than 3 levels under any content root.
But checking if the given mapping is valid doesn't require such detected: just check if Git directory is in the file system in the expected place.
Add a test, update the GitRootErrorsFinder.