[git] Introduce GitRemoteBranch, GitSvnRemoteBranch, GitLocalBranch
There is a significant difference between local and remote branches (the latter have 2 name formats and a link to a remote) => use separate classes for them.
git-svn has also different syntax.

TODO: GitRepositoryReader still initializes a GitBranch object, which is not fair, but the right initialization required pre-reading the list of remotes.