[git] reduce logging noise
This is actually a valid situation to request GitRepositoryManager#getRepositoryForFile for non-Git files. Can be used, for instance, in #update() of Git-only actions to check if the action should be visible at all (when it is under Svn, for instance).
(cherry picked from commit 32e7753a3829874764ac5d80abffc99a46174122)