[github] Fix connection timeout value. IDEA-101292
Set connection timeout params on the connection manager instead of
setting them on the client.

Setting these params on the client has no effect, because HttpConnection
is created in HttpMethodDirector#executeMethod, and it inherits params
from the HttpConnectionManager.

Connection parameters are applied to the HttpConnection from
HttpMethod (which inherits them from HttpClient) only AFTER connection
is open, see HttpMethodDirector#executeWithRetry (this.conn.open()
and only then applyConnectionParams()), but the SO_TIMEOUT is already
needed when _opening_ connection.