[tests] Create individual temp directory for each test run
To avoid problems when the previous test run is terminated abnormally
and therefore leaves the dirty temp dir.

Introduce UsefulTestCase#shouldDeleteTempFilesOnTearDown
and override it in GitCucumberLightProjectFixture instead of
overriding shouldContainTempFiles(). Thus temp directory with a random
suffix will be created for each test run, but will be deleted only
on application exit.