[git] IDEA-103375 Fix understanding Git submodule roots
Git submodule repository is located not in submodule/.git, but
in supermodule/modules/.git.
At the same time submodule/.git is a file which contains the path to
the real repository location.
We already detect this, but we didn't use refresh to search for the
root as a VirtualFile, and in most cases it was not in the VFS yet.
(cherry picked from commit 736c7606456e8af4effbc999820aaedaada6af12)