[git] IDEA-104846 Fix pushing new branch without new commits
The bug was introduce in b200615c when we stopped showing last 5
commits for new branches, and began to show only those commits which
will be actually pushed.

To avoid the problem check for "nothing to push" (0 commits) only in
the standard case, when the branch is pushed to its tracked branch.
(cherry picked from commit 9dadba3a2ab286f04317116f2d9ad549d4e9eb54)