[git] IDEA-105577 Revert changes in listing commits in Git Push dialog.
This reverts:
* bebc84c2 IDEA-104846 Fix pushing new branch without new commits
* b200615c for new branch show commits which are not pushed yet

The right fix for this issue would be to identify probable remote for
the new branch before showing push dialog, but this change may be too
dangerous for 129, so just reverting the changes which led to the
problem.

Reverting it in master as well, since it will be fixed within push
dialog redesign.

(cherry picked from commit 025340e1ae8556745b7697e9e890c0a11c22ee4f)