[git] IDEA-102290 Fix south panel for Untracked files dialog
The bug was caused by the fact that SelectFilesDialog called init()
in its constructor, therefore the descendants were unable to
customize the dialog by passing data via instance fields from the
constructor, because, for example, getSouthPanel() was already called
in super().

To solve this make SelectFilesDialog.init() for those who need just
to use the dialog, and don't call init() from SelectFilesDialog()
for those who need to extend it.