[vcs] IDEA-105707 Update Project Info: fix show diff for Created files
The root cause was that the file type of the Created change was
identified incorrectly as "Modified" and treated further as modified
(and it was not possible to show this modified file without the
"before"-content).

Fix looking for the closest parent GroupTreeNode: when changes are
grouped by folders, the GroupTreeNode is not the first parent,
but the first parent after all directory nodes.
(cherry picked from commit b21c42a80d491e75c5467ea6d99d3207ca532f63)