[git] Rename GitCommit -> GitHeavyCommit as a preparation to deprecation
GitCommit object is cluttered with a lot of fields which sometimes
are populated, sometimes not, and some of which are completely
unrelated to the commit object (like tags, branches or current branch).

It should be removed and changed to something more clear,
but the name is good => renaming the old bad guy.